# vulkan: optimizations for direct convolution