@jeffbolznv 2025-07-29
Please share the command line you used for sd. Maybe this is the same as the 4070 regression, I'll look into it.